TRNG_PIYUSH/requirements.txt
asttokens==3.0.1
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cupy-cuda12x==13.6.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fastrlock==0.8.3
fonttools==4.60.1
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
matplotlib==3.10.7
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
pure_eval==0.2.3
pycryptodome==3.23.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
reportlab==4.4.5
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.14
